show error messages when clicking wrong element
this additionally fixes a bug: due to restriction of element movement only for elements of the current interaction, the elements did not get an initial start value when starting the programm. Now the movement functions of faders, knobs and buttons use an additional flag to determine wether the movement is intial or not.